- Properly implemented the camera matrix (taken at WorldBegin) - Now taking into account frame aspect ration when building projection matrix - The teapot now renders, though somewhat outside the screen (why ?!)

+++ RI_System/include/RI_Options.h
@@ -39,7 +39,7 @@ public:
 
 	// Camera
 	int		mXRes, mYRes;
-	int		mPixelAspectRatio;
+	float	mPixelAspectRatio;
 	float	mFrameAspectRatio;
 	float	mLeft, mRight, mBottom, mTop;
 	float	mXMin, mXMax, mYMin, mYMax;
